Fix multiple requests to ampie on SPAs
When you open the info bar on an SPA like twitter.com and then navigate
to some other url on the domain, ampie sends get-links-pages-info every
time the url changes, even though the info bar is not open. I am not
sure why that happens exactly (the query should happen only when the
adjacent-links component is created for the first time), but added a
fix.